workbenchThemeService: move more to common
