+++ src/vs/workbench/services/themes/common/themeCompatibility.ts
